Add error handling module
Error handling is an essential part of a good API, and the RμSH Library
is no exception. An earlier experiment had been made to incorporate
exception handling with the use of setjmp() and longjmp(), but this did
not scale well. As a consequence, we're reverting to using the
traditional approach of returning error codes from functions, with
additional macros to serve as syntactic sugar.

The error handling module has been defined with the following interface:
  - `rush_erno` type
  - `RUSH_ERNO_NUL` symbolic constant
  - `RUSH_ERNO_HANDLE` symbolic constant
  - `RUSH_ERNO_MPOOL` symbolic constant
  - `RUSH_ERNO_RANGE` symbolic constant
  - `RUSH_TRY` label
  - `RUSH_CATCH` label
  - `RUSH_FINALLY` label
  - `rush_erno_get()` macro
  - `rush_erno_set()` macro
  - `rush_try()` macro
  - `rush_assert()` macro
  - `rush_assert_handle()` macro
  - `rush_assert_range()` macro

The `rush_mpool` implementation has also been refactored to make use of
the newly incorporated error handling module. Testing and documentation
are still pending.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>